Update README.md
This assignment focuses on the essentials of technical writing, teaching you how to create clear, concise, and user-friendly content. You’ll learn to tailor your writing for different audiences, simplify technical jargon, and use visuals and structure to enhance readability. By practicing these skills, you’ll be able to write effective documentation, tutorials, and Readme files that are easy to understand. The goal is to ensure your content is accessible and actionable, whether for beginners or experts, making complex information approachable for everyone.